requirements.txt
fastapi
uvicorn
stripe
twilio
firebase-admin
python-dotenv
reportlab
gunicorn
pillow
